Update testing convention, add husky precommit for format/linting consistency
