+++ .husky/pre-commit
@@ -1,0 +1,2 @@
+#!/bin/sh
+npx lint-staged
