+++ src/config/manager.ts
@@ -1,7 +1,9 @@
 import { z } from 'zod';
 import { logger } from '../utils/logger.js';
 import { llmConfigSchema } from './schemas.js';
-import type { AgentConfig, CLIConfigOverrides, LLMConfig, LLMProvenance } from './types.js';
+import type { AgentConfig, CLIConfigOverrides, LLMProvenance } from './types.js';
+
+declare function structuredClone<T>(value: T): T;
 
 /**
  * ConfigManager encapsulates merging file-based configuration, CLI overrides,
@@ -16,9 +18,11 @@ export class ConfigManager {
     private provenance: { llm: LLMProvenance };
 
     constructor(fileConfig: AgentConfig) {
+        // Note: structuredClone requires Node.js v17.0.0+. For older versions, use a polyfill or lodash.cloneDeep.
         this.resolved = structuredClone(fileConfig);
         this.provenance = { llm: { provider: 'file', model: 'file', router: 'default' } };
         this.applyDefaults();
+        this.validate(); // Fail fast on invalid fileConfig
     }
 
     private applyDefaults() {
@@ -47,6 +51,9 @@ export class ConfigManager {
             this.resolved.llm.apiKey = cliArgs.apiKey;
             // Optionally track provenance for apiKey if desired
         }
+        // Re-apply defaults in case some overrides were incorrect
+        this.applyDefaults();
+        this.validate(); // Fail fast on invalid overrides
         return this;
     }
