--- src/utils/tests/user-info.test.ts
@@ -1,7 +1,0 @@
-import { getUserId } from '../user-info.js';
-
-describe('getUserId', () => {
-  it('returns the default user id', () => {
-    expect(getUserId()).toBe('default-user');
-  });
-});
