+++ src/utils/user-info.test.ts
@@ -1,0 +1,8 @@
+import { describe, it, expect } from 'vitest';
+import { getUserId } from './user-info.js';
+
+describe('getUserId', () => {
+    it('returns the default user id', () => {
+        expect(getUserId()).toBe('default-user');
+    });
+});
