+++ vitest.config.ts
@@ -1,11 +1,11 @@
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
-  test: {
-    globals: true,
-    environment: 'node',
-    include: ['**/tests/**/*.test.ts', '**/tests/**/*.spec.ts'],
-    watch: true,
-    threads: true,
-  },
-}); 
+    test: {
+        globals: true,
+        environment: 'node',
+        include: ['**/*.test.ts', '**/*.spec.ts'],
+        watch: true,
+        threads: true,
+    },
+});
